+++ Source/Scene/GltfLoader.js
@@ -17,6 +17,7 @@ import InstanceAttributeSemantic from "./InstanceAttributeSemantic.js";
 import InterpolationType from "../Core/InterpolationType.js";
 import Matrix4 from "../Core/Matrix4.js";
 import ModelComponents from "./ModelComponents.js";
+import PrimitiveLoadPlan from "./PrimitiveLoadPlan.js";
 import numberOfComponentsForType from "./GltfPipeline/numberOfComponentsForType.js";
 import Quaternion from "../Core/Quaternion.js";
 import ResourceCache from "./ResourceCache.js";
@@ -49,14 +50,86 @@ const MetallicRoughness = ModelComponents.MetallicRoughness;
 const SpecularGlossiness = ModelComponents.SpecularGlossiness;
 const Material = ModelComponents.Material;
 
+/**
+ * States of the glTF loading process. These states also apply to
+ * asynchronous texture loading unless otherwise noted
+ *
+ * @enum {Number}
+ *
+ * @private
+ */
 const GltfLoaderState = {
+  /**
+   * The initial state of the glTF loader before load() is called.
+   *
+   * @type {Number}
+   * @constant
+   *
+   * @private
+   */
   UNLOADED: 0,
+  /**
+   * The state of the loader while waiting for the glTF JSON loader promise
+   * to resolve.
+   *
+   * @type {Number}
+   * @constant
+   *
+   * @private
+   */
   LOADING: 1,
+  /**
+   * The state of the loader once the glTF JSON is loaded but before
+   * process() is called.
+   *
+   * @type {Number}
+   * @constant
+   */
   LOADED: 2,
+  /**
+   * The state of the loader while parsing the glTF and creating GPU resources
+   * as needed.
+   *
+   * @type {Number}
+   * @constant
+   */
   PROCESSING: 3,
-  PROCESSED: 4,
-  READY: 4,
-  FAILED: 5,
+  /**
+   * For some features like handling CESIUM_primitive_outlines, the geometry
+   * must be modified after it is loaded. The post-processing state handles
+   * any geometry modification (if needed).
+   * <p>
+   * This state is not used for asynchronous texture loading.
+   * </p>
+   *
+   * @type {Number}
+   * @constant
+   */
+  POST_PROCESSING: 4,
+  /**
+   * Once the processing/post-processing states are finished, the loader
+   * enters the processed state (sometimes from a promise chain). The next
+   * call to process() will advance to the ready state.
+   *
+   * @type {Number}
+   * @constant
+   */
+  PROCESSED: 5,
+  /**
+   * When the loader reaches the ready state, the loaders' promise will be
+   * resolved.
+   *
+   * @type {Number}
+   * @constant
+   */
+  READY: 6,
+  /**
+   * If an error occurs at any point, the loader switches to the failed state.
+   *
+   * @type {Number}
+   * @constant
+   */
+  FAILED: 7,
 };
 
 /**
@@ -82,6 +155,7 @@ const GltfLoaderState = {
  * @param {Boolean} [options.loadAttributesAsTypedArray=false] Load all attributes and indices as typed arrays instead of GPU buffers.
  * @param {Boolean} [options.loadAttributesFor2D=false] If true, load the positions buffer and any instanced attribute buffers as typed arrays for accurately projecting models to 2D.
  * @param {Boolean} [options.loadIndicesForWireframe=false] If true, load the index buffer as both a buffer and typed array. The latter is useful for creating wireframe indices in WebGL1.
+ * @param {Boolean} [options.loadPrimitiveOutline=true] If true, load outlines from the {@link https://github.com/KhronosGroup/glTF/tree/master/extensions/2.0/Vendor/CESIUM_primitive_outline|CESIUM_primitive_outline} extension. This can be set false to avoid post-processing geometry at load time.
  * @param {Boolean} [options.renameBatchIdSemantic=false] If true, rename _BATCHID or BATCHID to _FEATURE_ID_0. This is used for .b3dm models
  * @private
  */
@@ -107,6 +181,8 @@ export default function GltfLoader(options) {
     options.loadIndicesForWireframe,
     false
   );
+  const loadPrimitiveOutline = defaultValue(options.loadPrimitiveOutline, true);
+
   const renameBatchIdSemantic = defaultValue(
     options.renameBatchIdSemantic,
     false
@@ -130,6 +206,7 @@ export default function GltfLoader(options) {
   this._loadAttributesAsTypedArray = loadAttributesAsTypedArray;
   this._loadAttributesFor2D = loadAttributesFor2D;
   this._loadIndicesForWireframe = loadIndicesForWireframe;
+  this._loadPrimitiveOutline = loadPrimitiveOutline;
   this._renameBatchIdSemantic = renameBatchIdSemantic;
 
   // When loading EXT_feature_metadata, the feature tables and textures
@@ -146,6 +223,11 @@ export default function GltfLoader(options) {
   this._process = function (loader, frameState) {};
   this._processTextures = function (loader, frameState) {};
 
+  // Information about whether to load primitives as typed arrays or buffers,
+  // and whether post-processing is needed after loading (e.g. for
+  // generating outlines)
+  this._primitiveLoadPlans = [];
+
   // Loaders that need to be processed before the glTF becomes ready
   this._loaderPromises = [];
   this._textureLoaders = [];
@@ -153,6 +235,11 @@ export default function GltfLoader(options) {
   this._bufferViewLoaders = [];
   this._geometryLoaders = [];
   this._structuralMetadataLoader = undefined;
+
+  // In some cases where geometry post-processing is needed (like generating
+  // outlines) new attributes are added that may have GPU resources attached.
+  // The GltfLoader will own the resources and store them here.
+  this._postProcessBuffers = [];
 
   // Loaded results
   this._components = undefined;
@@ -298,6 +385,11 @@ GltfLoader.prototype.load = function () {
           processLoaders(loader, frameState);
         }
 
+        if (loader._state === GltfLoaderState.POST_PROCESSING) {
+          postProcessGeometry(loader, frameState.context);
+          loader._state = GltfLoaderState.PROCESSED;
+        }
+
         if (loader._state === GltfLoaderState.PROCESSED) {
           unloadBufferViews(loader); // Buffer views can be unloaded after the data has been copied
           loader._state = GltfLoaderState.READY;
@@ -366,21 +458,66 @@ function handleError(gltfLoader, error) {
 }
 
 function processLoaders(loader, frameState) {
-  let i;
   const bufferViewLoaders = loader._bufferViewLoaders;
   const bufferViewLoadersLength = bufferViewLoaders.length;
-  for (i = 0; i < bufferViewLoadersLength; ++i) {
+  for (let i = 0; i < bufferViewLoadersLength; ++i) {
     bufferViewLoaders[i].process(frameState);
   }
 
   const geometryLoaders = loader._geometryLoaders;
   const geometryLoadersLength = geometryLoaders.length;
-  for (i = 0; i < geometryLoadersLength; ++i) {
+  for (let i = 0; i < geometryLoadersLength; ++i) {
     geometryLoaders[i].process(frameState);
   }
 
   if (defined(loader._structuralMetadataLoader)) {
     loader._structuralMetadataLoader.process(frameState);
+  }
+}
+
+function postProcessGeometry(loader, context) {
+  // Apply post-processing steps on geometry such as
+  // updating attributes for rendering outlines.
+  const loadPlans = loader._primitiveLoadPlans;
+  const length = loadPlans.length;
+  for (let i = 0; i < length; i++) {
+    const loadPlan = loadPlans[i];
+    loadPlan.postProcess(context);
+
+    if (loadPlan.needsOutlines) {
+      // The glTF loader takes ownership of any buffers generated in the
+      // post-process stage since they were created after the geometry loaders
+      // finished. This way they can be destroyed when the loader is destroyed.
+      gatherPostProcessBuffers(loader, loadPlan);
+    }
+  }
+}
+
+function gatherPostProcessBuffers(loader, primitiveLoadPlan) {
+  const buffers = loader._postProcessBuffers;
+  const primitive = primitiveLoadPlan.primitive;
+
+  const outlineCoordinates = primitive.outlineCoordinates;
+  if (defined(outlineCoordinates)) {
+    // outline coordinates are always loaded as a buffer.
+    buffers.push(outlineCoordinates.buffer);
+  }
+
+  // to do post-processing, all the attributes are loaded as typed arrays
+  // so if a buffer exists, it was newly generated
+  const attributes = primitive.attributes;
+  const length = attributes.length;
+  for (let i = 0; i < length; i++) {
+    const attribute = attributes[i];
+    if (defined(attribute.buffer)) {
+      buffers.push(attribute.buffer);
+    }
+  }
+
+  // Similarly for the indices.
+  const indices = primitive.indices;
+  if (defined(indices) && defined(indices.buffer)) {
+    buffers.push(indices.buffer);
   }
 }
 
@@ -432,16 +569,14 @@ function loadVertexBuffer(
   return vertexBufferLoader;
 }
 
-function loadIndexBuffer(loader, gltf, accessorId, draco, frameState) {
-  const loadAttributesAsTypedArray = loader._loadAttributesAsTypedArray;
-
-  // Load the index buffer as a typed array to generate wireframes in WebGL1.
-  const loadForWireframe =
-    loader._loadIndicesForWireframe && !frameState.context.webgl2;
-
-  const loadBuffer = !loadAttributesAsTypedArray;
-  const loadTypedArray = loadAttributesAsTypedArray || loadForWireframe;
-
+function loadIndexBuffer(
+  loader,
+  gltf,
+  accessorId,
+  draco,
+  loadBuffer,
+  loadTypedArray
+) {
   const indexBufferLoader = ResourceCache.loadIndexBuffer({
     gltf: gltf,
     accessorId: accessorId,
@@ -763,6 +898,7 @@ function loadVertexAttribute(
   gltfSemantic,
   draco,
   hasInstances,
+  needsPostProcessing,
   frameState
 ) {
   const semanticInfo = getSemanticInfo(
@@ -780,12 +916,26 @@ function loadVertexAttribute(
     loader._loadAttributesFor2D &&
     !frameState.scene3DOnly;
 
-  const loadAsTypedArrayOnly = loader._loadAttributesAsTypedArray;
-  const loadBuffer = !loadAsTypedArrayOnly;
-  const loadTypedArray = loadAsTypedArrayOnly || loadFor2D;
-  const loadAsTypedArrayPacked = false;
+  // Whether the final output should be a buffer or typed array
+  // after loading and post-processing.
+  const outputTypedArrayOnly = loader._loadAttributesAsTypedArray;
+  const outputBuffer = !outputTypedArrayOnly;
+  const outputTypedArray = outputTypedArrayOnly || loadFor2D;
+  const outputTypedArrayPacked = false;
 
-  return loadAttribute(
+  // Determine what to load right now:
+  //
+  // - If post-processing is needed, load a packed typed array for
+  //   further processing, and defer the buffer loading until later.
+  // - On the other hand, if post-processing is not needed,
+  //   set the load flags directly
+  const loadBuffer = needsPostProcessing ? false : outputBuffer;
+  const loadTypedArray = needsPostProcessing ? true : outputTypedArray;
+  const loadAsTypedArrayPacked = needsPostProcessing
+    ? true
+    : outputTypedArrayPacked;
+
+  const attribute = loadAttribute(
     loader,
     gltf,
     accessorId,
@@ -796,6 +946,13 @@ function loadVertexAttribute(
     loadTypedArray,
     loadAsTypedArrayPacked
   );
+
+  const attributePlan = new PrimitiveLoadPlan.AttributeLoadPlan(attribute);
+  attributePlan.loadBuffer = outputBuffer;
+  attributePlan.loadTypedArray = outputTypedArray;
+  attributePlan.loadPackedTypedArray = outputTypedArrayPacked;
+
+  return attributePlan;
 }
 
 function loadInstancedAttribute(
@@ -839,7 +996,14 @@ function loadInstancedAttribute(
   );
 }
 
-function loadIndices(loader, gltf, accessorId, draco, frameState) {
+function loadIndices(
+  loader,
+  gltf,
+  accessorId,
+  draco,
+  needsPostProcessing,
+  frameState
+) {
   const accessor = gltf.accessors[accessorId];
   const bufferViewId = accessor.bufferView;
 
@@ -850,12 +1014,32 @@ function loadIndices(loader, gltf, accessorId, draco, frameState) {
   const indices = new Indices();
   indices.count = accessor.count;
 
+  const loadAttributesAsTypedArray = loader._loadAttributesAsTypedArray;
+  // Load the index buffer as a typed array to generate wireframes in WebGL1.
+  const loadForWireframe =
+    loader._loadIndicesForWireframe && !frameState.context.webgl2;
+
+  // Whether the final output should be a buffer or typed array
+  // after loading and post-processing.
+  const outputBuffer = !loadAttributesAsTypedArray;
+  const outputTypedArray = loadAttributesAsTypedArray || loadForWireframe;
+
+  // Determine what to load right now:
+  //
+  // - If post-processing is needed, load a packed typed array for
+  //   further processing, and defer the buffer loading until later.
+  // - On the other hand, if post-processing is not needed, set the load
+  //   flags directly
+  const loadBuffer = needsPostProcessing ? false : outputBuffer;
+  const loadTypedArray = needsPostProcessing ? true : outputTypedArray;
+
   const indexBufferLoader = loadIndexBuffer(
     loader,
     gltf,
     accessorId,
     draco,
-    frameState
+    loadBuffer,
+    loadTypedArray
   );
 
   const promise = indexBufferLoader.promise.then(function (indexBufferLoader) {
@@ -871,7 +1055,11 @@ function loadIndices(loader, gltf, accessorId, draco, frameState) {
 
   loader._loaderPromises.push(promise);
 
-  return indices;
+  const indicesPlan = new PrimitiveLoadPlan.IndicesLoadPlan(indices);
+  indicesPlan.loadBuffer = outputBuffer;
+  indicesPlan.loadTypedArray = outputTypedArray;
+
+  return indicesPlan;
 }
 
 function loadTexture(
@@ -968,7 +1156,6 @@ function loadMaterial(loader, gltf, gltfMaterial, supportedImageFormats) {
     material.pbrSpecularGlossiness = pbrSpecularGlossiness;
   } else if (defined(pbrMetallicRoughness)) {
     const metallicRoughness = new MetallicRoughness();
-    material.metallicRoughness = metallicRoughness;
 
     if (defined(pbrMetallicRoughness.baseColorTexture)) {
       metallicRoughness.baseColorTexture = loadTexture(
@@ -992,7 +1179,7 @@ function loadMaterial(loader, gltf, gltfMaterial, supportedImageFormats) {
     );
     metallicRoughness.metallicFactor = pbrMetallicRoughness.metallicFactor;
     metallicRoughness.roughnessFactor = pbrMetallicRoughness.roughnessFactor;
-    material.pbrMetallicRoughness = pbrMetallicRoughness;
+    material.metallicRoughness = metallicRoughness;
   }
 
   // Top level textures
@@ -1160,16 +1347,39 @@ function loadFeatureIdTextureLegacy(
   return featureIdTexture;
 }
 
-function loadMorphTarget(loader, gltf, target) {
+function loadMorphTarget(
+  loader,
+  gltf,
+  target,
+  needsPostProcessing,
+  primitiveLoadPlan,
+  frameState
+) {
   const morphTarget = new MorphTarget();
+
+  // Don't pass in draco object since morph targets can't be draco compressed
+  const draco = undefined;
+  const hasInstances = false;
 
   for (const semantic in target) {
     if (target.hasOwnProperty(semantic)) {
       const accessorId = target[semantic];
-      morphTarget.attributes.push(
-        // Don't pass in draco object since morph targets can't be draco compressed
-        loadVertexAttribute(loader, gltf, accessorId, semantic, undefined)
+
+      const attributePlan = loadVertexAttribute(
+        loader,
+        gltf,
+        accessorId,
+        semantic,
+        draco,
+        hasInstances,
+        needsPostProcessing,
+        frameState
       );
+      morphTarget.attributes.push(attributePlan.attribute);
+
+      // The load plan doesn't need to distinguish morph target attributes from
+      // regular attributes
+      primitiveLoadPlan.attributePlans.push(attributePlan);
     }
   }
 
@@ -1185,6 +1395,8 @@ function loadPrimitive(
   frameState
 ) {
   const primitive = new Primitive();
+  const primitivePlan = new PrimitiveLoadPlan(primitive);
+  loader._primitiveLoadPlans.push(primitivePlan);
 
   const materialId = gltfPrimitive.material;
   if (defined(materialId)) {
@@ -1200,6 +1412,20 @@ function loadPrimitive(
     gltfPrimitive.extensions,
     defaultValue.EMPTY_OBJECT
   );
+
+  let needsPostProcessing = false;
+  const outlineExtension = extensions.CESIUM_primitive_outline;
+  if (loader._loadPrimitiveOutline && defined(outlineExtension)) {
+    needsPostProcessing = true;
+    primitivePlan.needsOutlines = true;
+    primitivePlan.outlineIndices = loadPrimitiveOutline(
+      loader,
+      gltf,
+      outlineExtension,
+      primitivePlan
+    );
+  }
+
   const draco = extensions.KHR_draco_mesh_compression;
 
   const attributes = gltfPrimitive.attributes;
@@ -1207,17 +1433,18 @@ function loadPrimitive(
     for (const semantic in attributes) {
       if (attributes.hasOwnProperty(semantic)) {
         const accessorId = attributes[semantic];
-        primitive.attributes.push(
-          loadVertexAttribute(
-            loader,
-            gltf,
-            accessorId,
-            semantic,
-            draco,
-            hasInstances,
-            frameState
-          )
+        const attributePlan = loadVertexAttribute(
+          loader,
+          gltf,
+          accessorId,
+          semantic,
+          draco,
+          hasInstances,
+          needsPostProcessing,
+          frameState
         );
+        primitivePlan.attributePlans.push(attributePlan);
+        primitive.attributes.push(attributePlan.attribute);
       }
     }
   }
@@ -1226,13 +1453,34 @@ function loadPrimitive(
   if (defined(targets)) {
     const targetsLength = targets.length;
     for (let i = 0; i < targetsLength; ++i) {
-      primitive.morphTargets.push(loadMorphTarget(loader, gltf, targets[i]));
+      primitive.morphTargets.push(
+        loadMorphTarget(
+          loader,
+          gltf,
+          targets[i],
+          needsPostProcessing,
+          primitivePlan,
+          frameState
+        )
+      );
     }
   }
 
   const indices = gltfPrimitive.indices;
   if (defined(indices)) {
-    primitive.indices = loadIndices(loader, gltf, indices, draco, frameState);
+    const indicesPlan = loadIndices(
+      loader,
+      gltf,
+      indices,
+      draco,
+      needsPostProcessing,
+      frameState
+    );
+
+    if (defined(indicesPlan)) {
+      primitivePlan.indicesPlan = indicesPlan;
+      primitive.indices = indicesPlan.indices;
+    }
   }
 
   // With the latest revision, feature IDs are defined in EXT_mesh_features
@@ -1273,6 +1521,12 @@ function loadPrimitive(
   primitive.primitiveType = gltfPrimitive.mode;
 
   return primitive;
+}
+
+function loadPrimitiveOutline(loader, gltf, outlineExtension) {
+  const accessorId = outlineExtension.indices;
+  const useQuaternion = false;
+  return loadAccessor(loader, gltf, accessorId, useQuaternion);
 }
 
 // For EXT_mesh_features
@@ -1984,7 +2238,7 @@ function parse(
       if (loader.isDestroyed()) {
         return;
       }
-      loader._state = GltfLoaderState.PROCESSED;
+      loader._state = GltfLoaderState.POST_PROCESSING;
     })
     .catch(rejectPromise);
 
@@ -1994,6 +2248,8 @@ function parse(
       if (loader.isDestroyed()) {
         return;
       }
+
+      // post processing only applies for geometry
       loader._textureState = GltfLoaderState.PROCESSED;
     })
     .catch(rejectTexturesPromise);
@@ -2026,6 +2282,18 @@ function unloadGeometry(loader) {
   loader._geometryLoaders.length = 0;
 }
 
+function unloadGeneratedAttributes(loader) {
+  const buffers = loader._postProcessBuffers;
+  const length = buffers.length;
+  for (let i = 0; i < length; i++) {
+    const buffer = buffers[i];
+    if (!buffer.isDestroyed()) {
+      buffer.destroy();
+    }
+  }
+  buffers.length = 0;
+}
+
 function unloadStructuralMetadata(loader) {
   if (defined(loader._structuralMetadataLoader)) {
     loader._structuralMetadataLoader.destroy();
@@ -2046,6 +2314,7 @@ GltfLoader.prototype.unload = function () {
   unloadTextures(this);
   unloadBufferViews(this);
   unloadGeometry(this);
+  unloadGeneratedAttributes(this);
   unloadStructuralMetadata(this);
 
   this._components = undefined;