+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ### 1.96 - 2022-08-01
 
+##### Additions :tada:
+
+- Models and tilesets that use the `CESIUM_primitive_outline` extension can now toggle outlines at runtime with the `showOutline` property. Furthermore, the color of the outlines can now be controlled by the `outlineColor` property. [#10506](https://github.com/CesiumGS/cesium/pull/10506)
+
 ##### Fixes :wrench:
 
 - Fixed a crash in the 3D Tiles Feature Styling sandcastle that occurred when using `ModelExperimental`. [#10514](https://github.com/CesiumGS/cesium/pull/10514)